Conflicts:
	Gruntfile.js
	lib/salve/formats.js

+++ test/salve-convert.js
@@ -16,24 +16,42 @@ describe("salve-convert", function () {
     var outpath = ".tmp_rng_to_js_test";
 
     afterEach(function () {
-        fs.unlinkSync(outpath);
+        if (fs.exists(outpath))
+            fs.unlinkSync(outpath);
     });
 
-    function salve_convert(inpath, expath, params, done) {
-        var child = spawn("bin/salve-convert", params.concat([inpath, outpath]),
-                          {stdio: "inherit"});
+    function salve_convert(inpath, exp, params, done, exp_status) {
+        if (exp_status === undefined)
+            exp_status = 0;
+
+        var child = spawn("build/dist/bin/salve-convert",
+                          params.concat([inpath, outpath]),
+                          (!exp_status) ? {stdio: "inherit"} :
+                          {stdio: ["ignore", 1, "pipe"]});
+
+        var stderr = [];
+        if (exp_status) {
+            child.stderr.on('data', function (data) {
+                stderr.push(data);
+            });
+        }
+
         child.on('exit', function (code, signal) {
-            assert.equal(code, 0, "salve-convert exit status");
-            if (expath) {
+            assert.equal(code, exp_status, "salve-convert exit status");
+            if (!exp_status && exp) {
                 // The actual output from diff would not be that useful here.
-                spawn("diff", [outpath, expath], {stdio: 'ignore'})
+                spawn("diff", [outpath, exp], {stdio: 'ignore'})
                 .on('exit', function (code, signal) {
                     assert.equal(code, 0, "there was a difference");
                     done();
                 });
             }
-            else
+            else {
+                if (exp_status)
+                    assert.equal(stderr.join(""),
+                                 fs.readFileSync(exp).toString());
                 done();
+            }
         });
     }
 
@@ -41,10 +59,19 @@ describe("salve-convert", function () {
     var tests = fs.readdirSync(dir);
 
     tests.forEach(function (t) {
-        if (t.slice(-4) === ".rng") {
+        if (t.slice(-4) !== ".rng")
+            return;
+
+        if (t.lastIndexOf("fails", 0) === -1) {
             var expected = t.slice(0, -4) + ".js";
             it("convert " + t, function (done) {
                 salve_convert(dir + t, dir + expected, [], done);
+            });
+        }
+        else {
+            it("convert fails on " + t, function (done) {
+                salve_convert(dir + t, dir + t.slice(0, -4) + ".txt",
+                              ["--include-paths"], done, 1);
             });
         }
     });
@@ -54,6 +81,7 @@ describe("salve-convert", function () {
         var expath = "test/tei/simplified-rng-v1.js";
 
         salve_convert(inpath, expath, ["--simplified-input",
+                                       "--allow-incomplete-types=quiet",
                                        "--no-optimize-ids"], done);
     });
 
@@ -61,14 +89,16 @@ describe("salve-convert", function () {
         var inpath = "test/tei/simplified.rng";
         var expath = "test/tei/simplified-rng-v1-optimized-ids.js";
 
-        salve_convert(inpath, expath, ["--simplified-input"], done);
+        salve_convert(inpath, expath, ["--simplified-input",
+                                       "--allow-incomplete-types=quiet",
+                                      ], done);
     });
 
     it("default execution", function (done) {
         var inpath = "test/tei/myTEI.rng";
         var expath = "test/tei/simplified-rng-v1-optimized-ids.js";
 
-        salve_convert(inpath, expath, [], done);
+        salve_convert(inpath, expath, ["--allow-incomplete-types=quiet"], done);
     });
 
     it("include paths", function (done) {
@@ -77,6 +107,7 @@ describe("salve-convert", function () {
 
         // Test created to deal with an internal error, so we don't
         // check the output.
-        salve_convert(inpath, null, ["--include-paths"], done);
+        salve_convert(inpath, null, ["--allow-incomplete-types=quiet",
+                                     "--include-paths"], done);
     });
 });
